added prettier config, formatted files

+++ contracts/LockedCreation.sol
@@ -11,9 +11,12 @@ contract LockedCreation is LockedPoolz {
         uint256 _FinishTime, //Until what time the pool will end
         uint256 _StartAmount, //Total amount of the tokens to sell in the pool
         address _Owner // Who the tokens belong to
-    ) external payable notZeroAddress(_Owner) returns(uint256) {
+    ) external payable notZeroAddress(_Owner) returns (uint256) {
         TransferInToken(_Token, msg.sender, _StartAmount);
-        if(WhiteList_Address != address(0) && !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))){
+        if (
+            WhiteList_Address != address(0) &&
+            !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))
+        ) {
             PayFee(Fee);
         }
         CreatePool(_Token, _StartTime, _FinishTime, _StartAmount, _Owner);
@@ -25,7 +28,9 @@ contract LockedCreation is LockedPoolz {
         uint256[] calldata _FinishTime,
         uint256[] calldata _StartAmount,
         address[] calldata _Owner
-    )   external payable
+    )
+        external
+        payable
         isGreaterThanZero(_Owner.length)
         isBelowLimit(_Owner.length)
     {
@@ -33,12 +38,21 @@ contract LockedCreation is LockedPoolz {
         require(_StartTime.length == _FinishTime.length, "Date Array Invalid");
         require(_Owner.length == _StartAmount.length, "Amount Array Invalid");
         TransferInToken(_Token, msg.sender, Array.getArraySum(_StartAmount));
-        if(WhiteList_Address != address(0) && !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))){
+        if (
+            WhiteList_Address != address(0) &&
+            !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))
+        ) {
             PayFee(Fee * _Owner.length);
         }
         uint256 firstPoolId = Index;
-        for(uint i=0 ; i < _Owner.length; i++){
-            CreatePool(_Token, _StartTime[i], _FinishTime[i], _StartAmount[i], _Owner[i]);
+        for (uint256 i = 0; i < _Owner.length; i++) {
+            CreatePool(
+                _Token,
+                _StartTime[i],
+                _FinishTime[i],
+                _StartAmount[i],
+                _Owner[i]
+            );
         }
         uint256 lastPoolId = Index - 1;
         emit MassPoolsCreated(firstPoolId, lastPoolId);
@@ -51,20 +65,35 @@ contract LockedCreation is LockedPoolz {
         uint256[] calldata _FinishTime,
         uint256[] calldata _StartAmount,
         address[] calldata _Owner
-    )   external payable
+    )
+        external
+        payable
         isGreaterThanZero(_StartTime.length)
         isBelowLimit(_Owner.length * _FinishTime.length)
     {
         require(_Owner.length == _StartAmount.length, "Amount Array Invalid");
         require(_FinishTime.length == _StartTime.length, "Date Array Invalid");
-        TransferInToken(_Token, msg.sender, Array.getArraySum(_StartAmount) * _FinishTime.length);
+        TransferInToken(
+            _Token,
+            msg.sender,
+            Array.getArraySum(_StartAmount) * _FinishTime.length
+        );
         uint256 firstPoolId = Index;
-        if(WhiteList_Address != address(0) && !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))){
+        if (
+            WhiteList_Address != address(0) &&
+            !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))
+        ) {
             PayFee(Fee * _Owner.length * _FinishTime.length);
         }
-        for(uint i=0 ; i < _FinishTime.length ; i++){
-            for(uint j=0 ; j < _Owner.length ; j++){
-                CreatePool(_Token, _StartTime[i], _FinishTime[i], _StartAmount[j], _Owner[j]);
+        for (uint256 i = 0; i < _FinishTime.length; i++) {
+            for (uint256 j = 0; j < _Owner.length; j++) {
+                CreatePool(
+                    _Token,
+                    _StartTime[i],
+                    _FinishTime[i],
+                    _StartAmount[j],
+                    _Owner[j]
+                );
             }
         }
         uint256 lastPoolId = Index - 1;
